сделал мини-домашку №1

+++ sem1/SuvorovIA/mini_hw_1/task_1.cpp
@@ -1,16 +1,16 @@
 #include <iostream>
 
-float square_area()
+int square_area()
 {
-	int side = 0;
+  int side = 0;
 
-	std::cout << "Input your side: ";
-	std::cin >> side;
+  std::cout << "Input your side: ";
+  std::cin >> side;
 
-	return side * side;
+  return side * side;
 }
 
 int main()
 {
-	std::cout << square_area() << std::endl;
+  std::cout << square_area() << std::endl;
 }
